+++ packages/superclass_generator/example/lib/example.freezed.dart
@@ -24,8 +24,12 @@ mixin _$User {
   int? get age => throw _privateConstructorUsedError;
   Profile get profile => throw _privateConstructorUsedError;
 
+  /// Serializes this User to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
+
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $UserCopyWith<User> get copyWith => throw _privateConstructorUsedError;
 }
 
@@ -49,6 +53,8 @@ class _$UserCopyWithImpl<$Res, $Val extends User>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -72,6 +78,8 @@ class _$UserCopyWithImpl<$Res, $Val extends User>
     ) as $Val);
   }
 
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
   @override
   @pragma('vm:prefer-inline')
   $ProfileCopyWith<$Res> get profile {
@@ -101,6 +109,8 @@ class __$$UserImplCopyWithImpl<$Res>
   __$$UserImplCopyWithImpl(_$UserImpl _value, $Res Function(_$UserImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -155,11 +165,13 @@ class _$UserImpl implements _User {
             (identical(other.profile, profile) || other.profile == profile));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType, name, age, profile);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
   _$$UserImplCopyWith<_$UserImpl> get copyWith =>
@@ -187,8 +199,11 @@ abstract class _User implements User {
   int? get age;
   @override
   Profile get profile;
+
+  /// Create a copy of User
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   _$$UserImplCopyWith<_$UserImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -201,8 +216,12 @@ Profile _$ProfileFromJson(Map<String, dynamic> json) {
 mixin _$Profile {
   String get bio => throw _privateConstructorUsedError;
 
+  /// Serializes this Profile to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
+
+  /// Create a copy of Profile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $ProfileCopyWith<Profile> get copyWith => throw _privateConstructorUsedError;
 }
 
@@ -224,6 +243,8 @@ class _$ProfileCopyWithImpl<$Res, $Val extends Profile>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of Profile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -256,6 +277,8 @@ class __$$ProfileImplCopyWithImpl<$Res>
       _$ProfileImpl _value, $Res Function(_$ProfileImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of Profile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -294,11 +317,13 @@ class _$ProfileImpl implements _Profile {
             (identical(other.bio, bio) || other.bio == bio));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType, bio);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of Profile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
   _$$ProfileImplCopyWith<_$ProfileImpl> get copyWith =>
@@ -319,8 +344,11 @@ abstract class _Profile implements Profile {
 
   @override
   String get bio;
+
+  /// Create a copy of Profile
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   _$$ProfileImplCopyWith<_$ProfileImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -335,8 +363,12 @@ mixin _$$UserProfile {
   int? get age => throw _privateConstructorUsedError;
   Profile? get profile => throw _privateConstructorUsedError;
 
+  /// Serializes this $UserProfile to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
+
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $$UserProfileCopyWith<$UserProfile> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -362,6 +394,8 @@ class _$$UserProfileCopyWithImpl<$Res, $Val extends $UserProfile>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -385,6 +419,8 @@ class _$$UserProfileCopyWithImpl<$Res, $Val extends $UserProfile>
     ) as $Val);
   }
 
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
   @override
   @pragma('vm:prefer-inline')
   $ProfileCopyWith<$Res>? get profile {
@@ -420,6 +456,8 @@ class __$$$UserProfileImplCopyWithImpl<$Res>
       _$$UserProfileImpl _value, $Res Function(_$$UserProfileImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -474,11 +512,13 @@ class _$$UserProfileImpl implements _$UserProfile {
             (identical(other.profile, profile) || other.profile == profile));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType, name, age, profile);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
   _$$$UserProfileImplCopyWith<_$$UserProfileImpl> get copyWith =>
@@ -507,8 +547,370 @@ abstract class _$UserProfile implements $UserProfile {
   int? get age;
   @override
   Profile? get profile;
+
+  /// Create a copy of $UserProfile
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   _$$$UserProfileImplCopyWith<_$$UserProfileImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+$UserRegistrationForm _$$UserRegistrationFormFromJson(
+    Map<String, dynamic> json) {
+  return _$UserRegistrationForm.fromJson(json);
+}
+
+/// @nodoc
+mixin _$$UserRegistrationForm {
+  String? get name => throw _privateConstructorUsedError;
+  String get street => throw _privateConstructorUsedError;
+  String get city => throw _privateConstructorUsedError;
+  String get country => throw _privateConstructorUsedError;
+
+  /// Serializes this $UserRegistrationForm to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of $UserRegistrationForm
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $$UserRegistrationFormCopyWith<$UserRegistrationForm> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $$UserRegistrationFormCopyWith<$Res> {
+  factory $$UserRegistrationFormCopyWith($UserRegistrationForm value,
+          $Res Function($UserRegistrationForm) then) =
+      _$$UserRegistrationFormCopyWithImpl<$Res, $UserRegistrationForm>;
+  @useResult
+  $Res call({String? name, String street, String city, String country});
+}
+
+/// @nodoc
+class _$$UserRegistrationFormCopyWithImpl<$Res,
+        $Val extends $UserRegistrationForm>
+    implements $$UserRegistrationFormCopyWith<$Res> {
+  _$$UserRegistrationFormCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of $UserRegistrationForm
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? name = freezed,
+    Object? street = null,
+    Object? city = null,
+    Object? country = null,
+  }) {
+    return _then(_value.copyWith(
+      name: freezed == name
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String?,
+      street: null == street
+          ? _value.street
+          : street // ignore: cast_nullable_to_non_nullable
+              as String,
+      city: null == city
+          ? _value.city
+          : city // ignore: cast_nullable_to_non_nullable
+              as String,
+      country: null == country
+          ? _value.country
+          : country // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$$UserRegistrationFormImplCopyWith<$Res>
+    implements $$UserRegistrationFormCopyWith<$Res> {
+  factory _$$$UserRegistrationFormImplCopyWith(
+          _$$UserRegistrationFormImpl value,
+          $Res Function(_$$UserRegistrationFormImpl) then) =
+      __$$$UserRegistrationFormImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String? name, String street, String city, String country});
+}
+
+/// @nodoc
+class __$$$UserRegistrationFormImplCopyWithImpl<$Res>
+    extends _$$UserRegistrationFormCopyWithImpl<$Res,
+        _$$UserRegistrationFormImpl>
+    implements _$$$UserRegistrationFormImplCopyWith<$Res> {
+  __$$$UserRegistrationFormImplCopyWithImpl(_$$UserRegistrationFormImpl _value,
+      $Res Function(_$$UserRegistrationFormImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of $UserRegistrationForm
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? name = freezed,
+    Object? street = null,
+    Object? city = null,
+    Object? country = null,
+  }) {
+    return _then(_$$UserRegistrationFormImpl(
+      name: freezed == name
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String?,
+      street: null == street
+          ? _value.street
+          : street // ignore: cast_nullable_to_non_nullable
+              as String,
+      city: null == city
+          ? _value.city
+          : city // ignore: cast_nullable_to_non_nullable
+              as String,
+      country: null == country
+          ? _value.country
+          : country // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$$UserRegistrationFormImpl implements _$UserRegistrationForm {
+  const _$$UserRegistrationFormImpl(
+      {this.name,
+      required this.street,
+      required this.city,
+      required this.country});
+
+  factory _$$UserRegistrationFormImpl.fromJson(Map<String, dynamic> json) =>
+      _$$$UserRegistrationFormImplFromJson(json);
+
+  @override
+  final String? name;
+  @override
+  final String street;
+  @override
+  final String city;
+  @override
+  final String country;
+
+  @override
+  String toString() {
+    return '\$UserRegistrationForm(name: $name, street: $street, city: $city, country: $country)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$$UserRegistrationFormImpl &&
+            (identical(other.name, name) || other.name == name) &&
+            (identical(other.street, street) || other.street == street) &&
+            (identical(other.city, city) || other.city == city) &&
+            (identical(other.country, country) || other.country == country));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(runtimeType, name, street, city, country);
+
+  /// Create a copy of $UserRegistrationForm
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$$UserRegistrationFormImplCopyWith<_$$UserRegistrationFormImpl>
+      get copyWith => __$$$UserRegistrationFormImplCopyWithImpl<
+          _$$UserRegistrationFormImpl>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$$UserRegistrationFormImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _$UserRegistrationForm implements $UserRegistrationForm {
+  const factory _$UserRegistrationForm(
+      {final String? name,
+      required final String street,
+      required final String city,
+      required final String country}) = _$$UserRegistrationFormImpl;
+
+  factory _$UserRegistrationForm.fromJson(Map<String, dynamic> json) =
+      _$$UserRegistrationFormImpl.fromJson;
+
+  @override
+  String? get name;
+  @override
+  String get street;
+  @override
+  String get city;
+  @override
+  String get country;
+
+  /// Create a copy of $UserRegistrationForm
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$$UserRegistrationFormImplCopyWith<_$$UserRegistrationFormImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+mixin _$$AnonymousUser {
+  String get userName => throw _privateConstructorUsedError;
+  int? get age => throw _privateConstructorUsedError;
+
+  /// Create a copy of $AnonymousUser
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $$AnonymousUserCopyWith<$AnonymousUser> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $$AnonymousUserCopyWith<$Res> {
+  factory $$AnonymousUserCopyWith(
+          $AnonymousUser value, $Res Function($AnonymousUser) then) =
+      _$$AnonymousUserCopyWithImpl<$Res, $AnonymousUser>;
+  @useResult
+  $Res call({String userName, int? age});
+}
+
+/// @nodoc
+class _$$AnonymousUserCopyWithImpl<$Res, $Val extends $AnonymousUser>
+    implements $$AnonymousUserCopyWith<$Res> {
+  _$$AnonymousUserCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of $AnonymousUser
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? userName = null,
+    Object? age = freezed,
+  }) {
+    return _then(_value.copyWith(
+      userName: null == userName
+          ? _value.userName
+          : userName // ignore: cast_nullable_to_non_nullable
+              as String,
+      age: freezed == age
+          ? _value.age
+          : age // ignore: cast_nullable_to_non_nullable
+              as int?,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$$AnonymousUserImplCopyWith<$Res>
+    implements $$AnonymousUserCopyWith<$Res> {
+  factory _$$$AnonymousUserImplCopyWith(_$$AnonymousUserImpl value,
+          $Res Function(_$$AnonymousUserImpl) then) =
+      __$$$AnonymousUserImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String userName, int? age});
+}
+
+/// @nodoc
+class __$$$AnonymousUserImplCopyWithImpl<$Res>
+    extends _$$AnonymousUserCopyWithImpl<$Res, _$$AnonymousUserImpl>
+    implements _$$$AnonymousUserImplCopyWith<$Res> {
+  __$$$AnonymousUserImplCopyWithImpl(
+      _$$AnonymousUserImpl _value, $Res Function(_$$AnonymousUserImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of $AnonymousUser
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? userName = null,
+    Object? age = freezed,
+  }) {
+    return _then(_$$AnonymousUserImpl(
+      userName: null == userName
+          ? _value.userName
+          : userName // ignore: cast_nullable_to_non_nullable
+              as String,
+      age: freezed == age
+          ? _value.age
+          : age // ignore: cast_nullable_to_non_nullable
+              as int?,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$$AnonymousUserImpl implements _$AnonymousUser {
+  const _$$AnonymousUserImpl({this.userName = 'Anonymous', this.age = 0});
+
+  @override
+  @JsonKey()
+  final String userName;
+  @override
+  @JsonKey()
+  final int? age;
+
+  @override
+  String toString() {
+    return '\$AnonymousUser(userName: $userName, age: $age)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$$AnonymousUserImpl &&
+            (identical(other.userName, userName) ||
+                other.userName == userName) &&
+            (identical(other.age, age) || other.age == age));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, userName, age);
+
+  /// Create a copy of $AnonymousUser
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$$AnonymousUserImplCopyWith<_$$AnonymousUserImpl> get copyWith =>
+      __$$$AnonymousUserImplCopyWithImpl<_$$AnonymousUserImpl>(
+          this, _$identity);
+}
+
+abstract class _$AnonymousUser implements $AnonymousUser {
+  const factory _$AnonymousUser({final String userName, final int? age}) =
+      _$$AnonymousUserImpl;
+
+  @override
+  String get userName;
+  @override
+  int? get age;
+
+  /// Create a copy of $AnonymousUser
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$$AnonymousUserImplCopyWith<_$$AnonymousUserImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }